+++ lib/furikake/resources/rds.rb
@@ -9,13 +9,14 @@ module Furikake
               {
                  subtitle: 'DB Instances',
                  header: ['DB Cluster Name', 'DB Instance Name',
-                          'DB Instance Class', 'DB Engine', 'DB Endpoint'],
+                          'DB Instance Class', 'DB Engine', 'DB Endpoint',
+                          'DB Instance Parameter Group', 'Security Group'],
                  resource: instance
               },
               {
                  subtitle: 'DB Clusters',
                  header: ['DB Cluster Name', 'Cluster Endpoint',
-                          'Cluster Reader Endpoint', 'Cluster Members'],
+                          'Cluster Reader Endpoint', 'DB Cluster Parameter Group', 'Cluster Members'],
                  resource: cluster
               }
           ]
@@ -34,6 +35,17 @@ module Furikake
           instance << i[:db_instance_class]
           instance << i[:engine]
           instance << i[:endpoint][:address]
+          instance << (i[:db_parameter_groups].map {|i| i[:db_parameter_group_name]}).join(',')
+
+          security_groups = []
+          if i[:db_security_groups].length > 0
+            security_groups << (i[:db_security_groups].map {|i| i[:db_security_group_name]})
+          end
+          if i[:vpc_security_groups].length > 0
+            security_groups << (i[:vpc_security_groups].map {|i| i[:vpc_security_group_id]})
+          end
+          instance << security_groups.sort.join('<br>')
+
           rds_infos << instance
         end
 
@@ -43,6 +55,7 @@ module Furikake
           cluster << c[:db_cluster_identifier]
           cluster << c[:endpoint]
           cluster << c[:reader_endpoint]
+          cluster << c[:db_cluster_parameter_group]
           cluster << (c[:db_cluster_members].map {|m| m[:is_cluster_writer] ? m[:db_instance_identifier] + '(W)' : m[:db_instance_identifier] + '(R)'}).join(', ')
           cluster_infos << cluster
         end